feat: move format method inside date adapter
