+++ src/global/config.ts
@@ -1,4 +1,4 @@
-import { documentLanguage } from './eventing';
+import { DateAdapter, NativeDateAdapter } from './datetime';
 
 export interface SbbIconConfig {
   interceptor?: (parameters: {
@@ -11,7 +11,7 @@ export interface SbbIconConfig {
 }
 
 export interface SbbDatetimeConfig {
-  format: (value: Date) => string;
+  dateAdapter: DateAdapter;
 }
 
 export interface SbbConfig {
@@ -22,27 +22,7 @@ export interface SbbConfig {
 export function readConfig(): SbbConfig {
   if (!('sbbConfig' in globalThis)) {
     globalThis.sbbConfig = {
-      datetime: {
-        format: (value: Date) => {
-          if (!value) {
-            return '';
-          }
-          const locale = documentLanguage() ? `${documentLanguage()}-CH` : 'en-CH';
-          const dateFormatter = new Intl.DateTimeFormat(locale, {
-            day: '2-digit',
-            month: '2-digit',
-            year: 'numeric',
-          });
-          const dayFormatter = new Intl.DateTimeFormat(locale, {
-            weekday: 'short',
-          });
-
-          let weekday = dayFormatter.format(value);
-          weekday = weekday.charAt(0).toUpperCase() + weekday.charAt(1);
-
-          return `${weekday}, ${dateFormatter.format(value)}`;
-        },
-      },
+      datetime: { dateAdapter: new NativeDateAdapter() },
     };
   }
   return globalThis.sbbConfig as SbbConfig;
